+++ HoloCollectionView/Classes/HoloMaker/Item/HoloCollectionViewItemMaker.h
@@ -1,0 +1,23 @@
+//
+//  HoloCollectionViewItemMaker.h
+//  HoloCollectionView
+//
+//  Created by 与佳期 on 2021/2/16.
+//
+
+#import <Foundation/Foundation.h>
+@class HoloCollectionItem, HoloCollectionItemMaker;
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface HoloCollectionViewItemMaker : NSObject
+
+@property (nonatomic, copy, readonly) HoloCollectionItemMaker *(^item)(Class item);
+
+@property (nonatomic, copy, readonly) HoloCollectionItemMaker *(^itemS)(NSString *itemString);
+
+- (NSArray<HoloCollectionItem *> *)install;
+
+@end
+
+NS_ASSUME_NONNULL_END